Update on "[DataLoader] Fix tempfile binding and removing for torch_shm_manager"
Fix the problem that `tempfile` has never been deleted even after `torch_shm_manager` is destroyed.
- The previous implementation has wrong path length for the Linux Socket. It leads to we lose the last character of the name of `tempfile` when bind the pathname to socket. At the end, we can not delete this file due to unexpected file name.
- After we solve the racing problem by introducing a temporary directory, it becomes more dangerous since it prevents `torch_shm_manager` to delete directory as the tempfile persists in the temporary directory.
- By having a sequence of clean up actions, we can make sure all resources are released. Then the temporary file and directory can be deleted.

Differential Revision: [D28202866](https://our.internmc.facebook.com/intern/diff/D28202866)

[ghstack-poisoned]

+++ torch/lib/libshm/socket.h
@@ -2,6 +2,7 @@
 
 #include <sys/types.h>
 #include <sys/socket.h>
+#include <sys/stat.h>
 #include <sys/un.h>
 #include <unistd.h>
 #include <poll.h>
@@ -28,7 +29,7 @@ protected:
 
   virtual ~Socket() {
     if (socket_fd != -1)
-      SYSCHECK_ERR_RETURN_NEG1(close(socket_fd));
+      close(socket_fd);
   }
 
   struct sockaddr_un prepare_address(const char *path) {
@@ -112,8 +113,10 @@ public:
     }
   }
 
-  virtual ~ManagerServerSocket() {
-    SYSCHECK_ERR_RETURN_NEG1(unlink(socket_path.c_str()));
+  void remove() {
+    struct stat st;
+    if (stat(socket_path.c_str(), &st) == 0)
+      SYSCHECK_ERR_RETURN_NEG1(unlink(socket_path.c_str()));
   }
 
   ManagerSocket accept() {
